gh-133 Refactor code location
