+++ docker/gaffer-jhub-options-server/src/templating.js
